fix-fe: ChatBox 원래대로 롤백
- ChatBox를 소켓 메세지, 이전 메세지 컴포넌트로 나눴었는데
새로운 메세지와 전의 메세지의 날짜가 따로 따로 생성되어 롤백
